scripts: gcloudnnc: Automate after reboot
The commands after rebooting, do not need to be inputted manually. External IP can be found by pinging a website and saved as a variable. Just bash script again and only copy the address outputted

+++ gcloudvnc.sh
@@ -13,44 +13,38 @@
 # • View your VNC instance from any device; viewing the progress of a proccess mobile (e.g. compiling, uploading etc) or 
 # to make slight changes that you don't need to be infront of a PC for.
 #
-# You could coincide this script with './buildrom.sh' if you wish, but it's mainly to provide a User Friendly
-# aspect to the Google Cloud Console; making best of the free trial credits, with the blessed specs and Internet
-# speeds Google has to offer.
-#
-sudo apt update && sudo apt upgrade -y
-# Update inastance's repository to be able to fetch and install all needed packages in next command.
-sudo apt install -y ubuntu-desktop gnome-panel gnome-settings-daemon metacity nautilus gnome-terminal vnc4server
-# Setup a GUI Desktop Environment.
-sudo ufw allow 5901:5910/tcp
-# Setup a Firewall to allow VNC access
-vncserver
-# Start up a vncserver
-#
-# User is prompted to enter a Password for the vncserver
-#
-# Password: *******
-# Verify: *******
-#
-sed -i /^x-terminal-emulator/d ~/.vnc/xstartup
-sed -i /^x-window-manager/d ~/.vnc/xstartup
-# Remove current environment assets; limited for terminal.
-echo -e 'gnome-panel &\ngnome-settings-daemon &\nmetacity &\nnautilus &\ngnome-terminal &' >> ~/.vnc/xstartup
-# Include the packages installed above as part of the new environment on VNC startup.
-sudo reboot
-# Let the changes take effect with a reboot, will disconnect the SSH.
-#
-# Upon re-opening the SSH Window:
-#$ vncserver
-# To run the vncserver for the final time.
-#
-#$ nc 'externalIPAddressOfInstance' 5901
-# This is to provide a 'n'ew 'c'onnection for the VNC Server.
-# You can find the External IP Address, next to the name you have given to your instance.
-#
-# You can then use any VNC viewer to view your GUI Enivronment, for example:
-# https://www.realvnc.com/en/connect/download/viewer/
-# In the 'Enter a VNC Address..' bar, based on the previous terminal command, type:
-# 'externalIPAddressOfInstance':5901
-#
-# Profit!
-#
+# This is mainly to provide a User Friendly aspect to the Google Cloud Console; making best of the free trial credits, with 
+# the blessed specs and Internet speeds Google has to offer.
+
+extip=$(curl ifconfig.me)
+
+if [[ ! -e gcloudvnc ]]; then
+	touch gcloudvnc
+	# Leave breadcrumb to show script has been bashed
+	sudo apt update && sudo apt upgrade -y
+	# Update inastance's repository to be able to fetch and install all needed packages in next command
+	sudo apt install -y ubuntu-desktop gnome-panel gnome-settings-daemon metacity nautilus gnome-terminal vnc4server
+	# Setup a GUI Desktop Environment.
+	sudo ufw allow 5901:5910/tcp
+	# Setup a Firewall to allow VNC access
+	vncserver
+	# Start up the vncserver
+	#
+	# User is prompted to enter a Password for the vncserver
+	#
+	sed -i /^x-terminal-emulator/d ~/.vnc/xstartup
+	sed -i /^x-window-manager/d ~/.vnc/xstartup
+	# Remove current environment assets; limited for terminal.
+	echo -e 'gnome-panel &\ngnome-settings-daemon &\nmetacity &\nnautilus &\ngnome-terminal &' >> ~/.vnc/xstartup
+	# Include the packages installed above as part of the new environment on VNC startup.
+	sudo reboot
+	# Let changes take effect.
+else
+	vncserver
+	# Start vncserver
+	# Use any VNC viewer to view GUI Enivronment. Reccommended: 'VNC Viewer': https://www.realvnc.com/en/connect/download/viewer/
+	printf "\n\n"$extip":5901\n\n"
+	# Input above address into VNC Viewer search bar
+	nc "$extip" 5901
+	# 'N'ew 'c'onntection to VNC Server at this address
+fi
